Update all calls to wait() to be timed calls.
jwh1

+++ trunk/TAO/examples/POA/Default_Servant/run_test.pl
@@ -12,14 +12,29 @@ $iorfile = "server.ior";
 
 unlink $iorfile;
 
-$SV = Process::Create ($EXEPREFIX."server$Process::EXE_EXT", " -o $iorfile");
+$SV = Process::Create ($EXEPREFIX."server".$Process::EXE_EXT, " -o $iorfile");
 
-ACE::waitforfile ($iorfile);
+if (ACE::waitforfile_timed ($iorfile, 5) == -1) {
+  print STDERR "ERROR: cannot find file <$iorfile>\n";
+  $SV->Kill (); $SV->TimedWait (1);
+  exit 1;
+}
 
-$status  = system ($EXEPREFIX."client$Process::EXE_EXT -k $iorfile");
+$CL  = Process::Create ($EXEPREFIX."client$Process::EXE_EXT ",
+			" -k $iorfile");
 
-$SV->Kill (); $SV->Wait ();
+$client = $CL->TimedWait (60);
+if ($client == -1) {
+  print STDERR "ERROR: client timedout\n";
+  $CL->Kill (); $CL->TimedWait (1);
+}
+
+$SV->Kill (); $SV->TimedWait (1);
 
 unlink $iorfile;
 
-exit $status;
+if ($client != 0) {
+  exit 1;
+}
+
+exit 0;
